+++ app/src/main/java/com/freshdigitable/udonroad/MainActivity.java
@@ -46,6 +46,7 @@ import com.freshdigitable.udonroad.listitem.OnUserIconClickedListener;
 import com.freshdigitable.udonroad.oauth.OAuthActivity;
 import com.freshdigitable.udonroad.repository.ImageRepository;
 import com.freshdigitable.udonroad.subscriber.ConfigRequestWorker;
+import com.freshdigitable.udonroad.timeline.TimelineFragment;
 import com.freshdigitable.udonroad.user.UserInfoActivity;
 
 import javax.inject.Inject;
