+++ app/src/main/java/com/freshdigitable/udonroad/timeline/TimelineAnimator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016. Matsuda, Akihit (akihito104)
+ * Copyright (c) 2018. Matsuda, Akihit (akihito104)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.freshdigitable.udonroad;
+package com.freshdigitable.udonroad.timeline;
 
 import android.support.v4.view.ViewCompat;
 import android.support.v4.view.ViewPropertyAnimatorCompat;
@@ -22,6 +22,8 @@ import android.support.v4.view.ViewPropertyAnimatorListenerAdapter;
 import android.support.v7.widget.RecyclerView.ViewHolder;
 import android.support.v7.widget.SimpleItemAnimator;
 import android.view.View;
+
+import com.freshdigitable.udonroad.BuildConfig;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -35,7 +37,7 @@ public class TimelineAnimator extends SimpleItemAnimator {
   @SuppressWarnings("unused")
   private static final String TAG = TimelineAnimator.class.getSimpleName();
 
-  public TimelineAnimator() {
+  TimelineAnimator() {
     super();
     setSupportsChangeAnimations(false);
   }
